解除 statuscheck 脚本指令 @sc_tickleft 返回值的上限

+++ src/map/script.cpp
@@ -26195,7 +26195,7 @@ BUILDIN_FUNC(statuscheck) {
 	// 之所以实现 statuscheck / sc_check 完全出于兼容目的考虑
 	struct TimerData* timer = (struct TimerData*)get_timer(sd->sc.data[id]->timer);
 	t_tick tickleft = (timer ? DIFF_TICK(timer->tick, gettick()) : -1);
-	pc_setreg(sd, add_str("@sc_tickleft"), (int)tickleft);
+	pc_setreg(sd, add_str("@sc_tickleft"), tickleft);
 
 	script_pushint(st, 1);
 	return SCRIPT_CMD_SUCCESS;
